+++ CHANGES.rst
@@ -12,6 +12,7 @@ v3.1.0 (xxxx-xx-xx)
 - Add ``matches_property``.
 - Add ``start_case``.
 - Make callbacks support ``matches_property`` style as ``[key, value]``.
+- Make callbacks support shallow ``pluck`` style callbacks as ``[key]``.
 
 
 .. _changelog-v3.0.0: